refactor(controller): extract queue.ts's non-class parts into queue/
queue.ts is 2681 lines, but ~2100 of those are one class. Splitting a
class means either free functions taking the instance or a mixin, both
of which trade readability for a line count — so the class stays put and
what surrounds it moves out:

  types.ts     the shapes that flow through the queue
  pure.ts      side-effect-free helpers and pacing constants
  kinds.ts     the voice-kind registry the DJ recap reads
  voice-io.ts  handoff-file writes + the spoken-segment serialiser

voice-io.ts is the one that earns its own file rather than just being
lines moved: the handoff write chain (#140), the spoken-segment
serialiser (#310) and the jingle collision guard (#997) are one
subsystem with a narrow interface, and they were interleaved with pure
comparison helpers at the bottom of the file.

queue.ts 2681 -> 2215, and it re-exports what it moved, so callers are
unchanged. Also un-exported 48 symbols across settings/, library-db/ and
queue/ that the earlier splits had made public purely as a side effect of
the extraction — nothing outside their own module referenced them.

Breaking the class up properly is worth doing, but as its own change with
its own review.

No behaviour change. Controller lint clean, all 52 test files pass.
